Erased "Cancelar cuenta button".
Now user have to climb the highest mountain in Westeros, kill a dragon (physically and intellectually), get its treasure, find one silver key in it and open the Final box. Inside of the box, if user is smart enough, it  will found the Final Envelope. Then he has to send us a certified card (via pidgeon_based mail) with 7 stamps of the 7 kingdoms.
Then, and only after carefoully revisied the stamps and the pidgeon, user will cancel the account.